[plot] Make endpoints have f32s instead of usizes

+++ src/aflak_plot/src/interactions.rs
@@ -62,7 +62,7 @@ pub struct FinedGrainedROI {
 #[derive(Clone, Debug, PartialEq)]
 pub struct Line {
     pub(crate) id: usize,
-    pub endpoints: ((usize, usize), (usize, usize)),
+    pub endpoints: ((f32, f32), (f32, f32)),
     pub endpointsfill: (bool, bool),
     pub pixels: Vec<(usize, usize)>,
 }
@@ -95,7 +95,7 @@ impl Line {
     pub fn new(id: usize) -> Self {
         Self {
             id,
-            endpoints: ((0, 0), (0, 0)),
+            endpoints: ((0.0, 0.0), (0.0, 0.0)),
             endpointsfill: (false, false),
             pixels: vec![],
         }
